requirements.txt
streamlit
pandas
matplotlib
supabase
google-generativeai
groq